api/requirements.txt
asgiref==3.11.1
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.0
et_xmlfile==2.0.0
greenlet==3.3.2
numpy==2.4.3
openpyxl==3.1.5
pandas==3.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
psycopg2==2.9.11
pycparser==3.0
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.48
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
